requirements.txt
einops==0.6.1
fastapi==0.111.0
ftfy==6.1.1
huggingface-hub==0.17.3
imageio==2.31.1
kornia==0.6.12
langchain==0.1.0
langgraph==0.5.3
lpips==0.1.4
matplotlib==3.8.1
natsort==8.3.1
omegaconf
opencv-python==4.8.1.78
pandas==2.1.1
Pillow==10.1.0
python-dotenv==1.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.2 --only-binary :all:
tqdm==4.66.1
transformers==4.34.0
wandb>=0.15.10